% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{calc.beta_cc}
\alias{calc.beta_cc}
\title{Time dependent factor beta_cc}
\usage{
calc.beta_cc(t, tref, sC)
}
\arguments{
\item{t}{time (days)}

\item{tref}{reference time (days)}

\item{sC}{factor according to table B.2 (EC2)}
}
\value{
residual strength f_Ri
}
\description{
This function returns time dependent factor beta_cc according to EC2 (B.2).
}
