% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{calc.charact.value}
\alias{calc.charact.value}
\title{Calculate characteristic value}
\usage{
calc.charact.value(x, Vknown = F, Vx = NA)
}
\arguments{
\item{x}{a vector containing numerical values of material properties}

\item{Vknown}{a boolean value T or F whether V is known (T) or not (F)}

\item{Vx}{known coefficient of variation}
}
\description{
This function returns characteristic value calculated according to EN1990, Annex D (D7).
}
