% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{calc.concr.prop.EC2}
\alias{calc.concr.prop.EC2}
\title{Calculate concrete properties}
\usage{
calc.concr.prop.EC2(x, conv_factor = 1)
}
\arguments{
\item{x}{a vector containing numerical values of concrete strength}

\item{conv_factor}{a transition factor depending on shape and size of the specimens}
}
\description{
This function returns dataframe of concrete properties calculated according to EN1992-1-1 (2023).
}
