% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{calc.Ecm.MC2020}
\alias{calc.Ecm.MC2020}
\title{Modulus of elasticity of concrete according to Model Code 2020}
\usage{
calc.Ecm.MC2020(fcm, alE = 1)
}
\arguments{
\item{fcm}{concrete compressive strength (mean value).}

\item{alE}{coefficient considering type of aggregates. Quartzite aggregates = 1.0, other 0.7...1.2}
}
\value{
modulus of elasticity Ecm
}
\description{
This function calculates modulus of elasticity of concrete according to Model Code 2020 (14.7-1).
}
