% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{calc.FRC.class}
\alias{calc.FRC.class}
\title{FRC post-cracking strength class}
\usage{
calc.FRC.class(fR1k, fR3k)
}
\arguments{
\item{fR1k}{residual strengths significant for serviceability.}

\item{fR3k}{residual strengths significant for ultimate conditions.}
}
\value{
A string representing FRC post-cracking strength class.
}
\description{
Returns FRC post-cracking strength class according to Model Code 2010.
}
