% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{calc.FRC.mat.model.LoAIII.1}
\alias{calc.FRC.mat.model.LoAIII.1}
\title{FRC material model in tension}
\usage{
calc.FRC.mat.model.LoAIII.1(
  fcm,
  fct = NA,
  fR1k,
  fR3k,
  wu = 2.5,
  al0 = 0.58/2,
  Act = 1,
  lcs = 125
)
}
\arguments{
\item{fR1k}{residual strengths significant for serviceability.}

\item{fR3k}{residual strengths significant for ultimate conditions.}
}
\value{
...
}
\description{
Returns FRC post-cracking strength class according to Model Code 2020.
}
