% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{calc.fRi}
\alias{calc.fRi}
\title{Residual strength according to EN 14651}
\usage{
calc.fRi(F, b, hsp, l)
}
\arguments{
\item{F}{load in three point bending test, kN}

\item{b}{width of the prism, mm}

\item{hsp}{height of the prism at the notch (from the tip of the notch), mm}

\item{l}{span of the prism (distance between the supports), mm}
}
\value{
residual strength f_Ri
}
\description{
This function returns residual tensile strength of notched prisms according to EN 14651.
}
