% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{calc.SIA162}
\alias{calc.SIA162}
\title{Calculations according to SIA 162/6}
\usage{
calc.SIA162(F_vect, w_vect, h0, lf, n)
}
\arguments{
\item{F_vect}{a vector containing force readings, kN}

\item{w_vect}{a vector containing displacement readings, mm}

\item{h0}{thickness of the slab, m}

\item{lf}{lengt of fibres, mm}

\item{n}{number of cracks}
}
\value{
A list of calculated values: fct, Gf
}
\description{
This function returns tensile strength and fracture energy calculated according to
the standard SIA 162/6, based on the round slabs.
}
