% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formTest.R
\name{cmod.long.to.wide}
\alias{cmod.long.to.wide}
\title{CMOD dataframe from long to wide format}
\usage{
cmod.long.to.wide(
  df,
  id_kopa,
  id_paraugi,
  id_CMOD,
  id_Load,
  cmod_vals = seq(0, 4, 0.01)
)
}
\arguments{
\item{df}{A long format data frame for Load - CMOD test results.}

\item{id_kopa}{name of the column corresponding to the sample (sērija/kopa) in long table.}

\item{id_paraugi}{name of the column containgin specimen numbers in long table.}

\item{id_CMOD}{name of the column containing CMOD readings in long table.}

\item{id_Load}{name of the column containing Load or residual strength values in long table.}

\item{cmod_vals}{vector of predefined CMOD values. If the first value is set to 0.05,
the max Load value in this range is calculated.}
}
\value{
A wide format data frame separating Load or residual strength values in separate columns for each specimen.
}
\description{
This function converts long format data frame to wide format.
This is needed for notched prism bending test results to calculate min, max and mean values
for different CMOD values. This function also calculates the min, max and mean values.
}
