% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{combine.two.df.time}
\alias{combine.two.df.time}
\title{Combines two measured values based on time}
\usage{
combine.two.df.time(dfRef, dfValue, time_diff = 0)
}
\arguments{
\item{dfRef}{a dataframe having two columns c("Time", "RefValue") in this order}

\item{dfValue}{a dataframe having two columns c("Time", "Value") in this order}

\item{time_diff}{time difference: dfValue$Time - dfRef$Time}
}
\value{
data frame with four columns: Time, Reference Value, Measured value, and Label
}
\description{
This function combines two data frames - both containing time and
a measurement (e.g. deflection, crack width, load, etc.).
It is used to make diagrams like Load vs deflection or Load vs crack width.
}
