% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{combine.values.with.load}
\alias{combine.values.with.load}
\title{Combines load with other measurements based on time}
\usage{
combine.values.with.load(dfLoad, dfValue, time_diff = 0)
}
\arguments{
\item{dfLoad}{a dataframe having two columns c("Time", "Load") in this order}

\item{dfValue}{a dataframe having two columns c("Time", "Value") in this order}

\item{time_diff}{time difference: dfValue$Time - dfLoad$Time}
}
\description{
This function combines two data frames - one containing time and load,
the other - time and another measurement (e.g. deflection, crack width).
It is used to make diagrams like Load vs deflection or Load vs crack width.
}
