% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{dataframe.wide.to.long}
\alias{dataframe.wide.to.long}
\title{Converts data frame from wide to long}
\usage{
dataframe.wide.to.long(df, x_value, y_value, repeated_columns = NULL)
}
\arguments{
\item{df}{A data frame to be converted}

\item{x_value}{A column name that contains values on x axis}

\item{y_value}{A vector of column names that contains values on y axis}

\item{repeated_columns}{A vector of column names that will be preserved and repeated for every label}
}
\value{
A converted data frame with tree columns
}
\description{
This function converts a data frame with multiple columns to long data frame,
where there are only three columns: x value, y value and label. It is done for
plotting purposes.
}
