% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{find.and.plot.peaks}
\alias{find.and.plot.peaks}
\title{Finds peaks and plots diagram with the peak values}
\usage{
find.and.plot.peaks(
  x1,
  x2,
  cols1,
  cols2,
  prob1 = 0.05,
  prob2 = 0.05,
  peak1 = 1,
  peak2 = 1
)
}
\arguments{
\item{x1, x2}{data frames containing measured data that needs to be compared}

\item{cols1, cols2}{column names of the data frames - c("Time", "Value") excpected
should be the same in x1 and peaks1 and in x2 and peaks2}

\item{prob1, prob2}{probability used as treshold to select the peak values}

\item{peak1, peak2}{numbers of peaks that represent the same time in the test (measuring process)}
}
\description{
This function finds peak values in measured data (two data frames), wich consists of time and value.
Then it plots the data in a diagram, pointing out the time at peak values.
}
