% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{find.peak}
\alias{find.peak}
\title{Finds peak values in measured data}
\usage{
find.peak(x, colTime, colValue, prob = 0.05)
}
\arguments{
\item{x}{a dataframe conteining atleast two columns (time and value)}

\item{colTime}{number or name of the column containing time}

\item{colValue}{number or name of the column containing measured values}

\item{prob}{probability used as treshold to select the peak values}
}
\description{
This function finds peak values in measured data, wich consists of time and value.
This function is used in function find.and.plot.peaks - to help to find time difference
caused by the time settings in different devices
}
