% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formTest.R
\name{formTestAll}
\alias{formTestAll}
\title{Form+Test test results}
\usage{
formTestAll(dir_path)
}
\arguments{
\item{dir_path}{A path to the folder where the files are stored.}
}
\value{
A data frame containing test results - raw data.
}
\description{
This function returns a data frame containing all the data obtained from files in a folder.
All the files must be of the same type, containing results of the same type of test procedure.
}
