% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formTest.R
\name{formTestASTMSumarry}
\alias{formTestASTMSumarry}
\title{Form+Test ASTM test results}
\usage{
formTestASTMSumarry(file)
}
\arguments{
\item{file}{A file name of a text file where the exported data are saved.}
}
\value{
A data frame containing summary of the results of ASTM round panel tests.
}
\description{
This function returns a data frame from a text file exported from Proteus software.
This function is similar to formTestCMODSumarry, but takes the summary data in case of ASTM round panel tests.
The data are in the same line with specimen number.
}
