% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formTest.R
\name{formTestCMODSumAll}
\alias{formTestCMODSumAll}
\title{Form+Test CMOD test results}
\usage{
formTestCMODSumAll(dir_path)
}
\arguments{
\item{dir_path}{A path to the folder where the files are stored.}
}
\value{
A data frame containing test results - summary of CMOD tests.
}
\description{
This function returns a data frame containing all the summary data of CMOS tests
obtained from files in a folder.
It is valid only with the files from CMOD standard test.
}
