% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formTest.R
\name{formTestCMODSumarry}
\alias{formTestCMODSumarry}
\title{Form+Test CMOD test results}
\usage{
formTestCMODSumarry(file)
}
\arguments{
\item{file}{A file name of a text file where the exported data are saved.}
}
\value{
A data frame containing summary of the results of CMOD tests.
}
\description{
This function returns a data frame from a text file exported from Proteus software.
This function is similar to formTestDataFrame, but takes only the summary data,
which are in the same line with specimen number
}
