% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formTest.R
\name{formTestDataFrame}
\alias{formTestDataFrame}
\title{Form+Test test results}
\usage{
formTestDataFrame(file)
}
\arguments{
\item{file}{A file name of a text file where the exported data are saved.}
}
\value{
A data frame containing test results - raw data.
}
\description{
This function returns a data frame from a text file exported from Proteus software.
In the file usualy results from one or more test specimens are stored.
}
