% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formTest.R
\name{formTestFlexSumarry}
\alias{formTestFlexSumarry}
\title{Form+Test prism bending test results}
\usage{
formTestFlexSumarry(file)
}
\arguments{
\item{file}{A file name of a text file where the exported data are saved.}
}
\value{
A data frame containing summary of the results of flexural tensile tests.
}
\description{
This function returns a data frame from a text file exported from Proteus software.
This function is similar to formTestCMODSumAll, but is used in plain concrete
flexural tensile tests according to EN 12390-5:2009
}
