% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formTest.R
\name{formTestSIASumAll}
\alias{formTestSIASumAll}
\title{Form+Test SIA162 test results}
\usage{
formTestSIASumAll(dir_path)
}
\arguments{
\item{dir_path}{A path to the folder where the files are stored.}
}
\value{
A data frame containing test results - summary of SIA162 tests.
}
\description{
This function returns a data frame containing all the summary data of SIA162 tests
obtained from files in a folder.
It is valid only with the files from SIA162 standard test.
}
