% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbm.R
\name{HBMDataFrame}
\alias{HBMDataFrame}
\title{HBM QuantumX data}
\usage{
HBMDataFrame(file)
}
\arguments{
\item{file}{A file name of a text file where the exported data are saved.}
}
\value{
A data frame containing test results - raw data.
}
\description{
This function returns a data frame from a text file exported from Catman Easy software.
In the file usualy results from one test measured by QuantumX 440B/1615B are stored.
}
