\name{LabData-package}
\alias{LabData-package}
\alias{LabData}
\docType{package}
\title{
\packageTitle{LabData}
}
\description{
\packageDescription{LabData}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LabData}
\packageIndices{LabData}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{LabData}

Maintainer: \packageMaintainer{LabData}
}
\references{
~~ Literature or other references for background information ~~
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
