% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matest.R
\name{matestCalcEmod}
\alias{matestCalcEmod}
\title{Calculate E modulus from Matest diagrams}
\usage{
matestCalcEmod(folder, load_levels)
}
\arguments{
\item{folder}{A folder path conteining txt files (E modulus) exported from Matest machine.}

\item{load_levels}{a vector consisting of two values defining load levels at which E modulus is calculated.}
}
\value{
A data frame containing test results.
}
\description{
This function returns data frame containing E modulus and Poisson's ratio,
taking form Matest load vs strain diagrams.
It assumes that each specimen is in a separate file and all the files ar in one folder.
}
