% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matest.R
\name{matestDataFrame}
\alias{matestDataFrame}
\title{Matest compression machine test results}
\usage{
matestDataFrame(file)
}
\arguments{
\item{file}{A file name of a text file where the exported data are saved.}
}
\value{
A data frame containing test results - raw data.
}
\description{
This function returns a data frame from a text file exported from Matest compression machine.
In the file usualy results from one test specimens are stored. This function reads only 4 columns,
which is the case if only force, piston displacement and time is measured.
}
