% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matest.R
\name{matestResult}
\alias{matestResult}
\title{Matest compression machine test results}
\usage{
matestResult(file)
}
\arguments{
\item{file}{A file name of a text file where the test data are saved.}
}
\value{
A data frame containing test results - raw data.
}
\description{
This function returns data of a tested specimen such as:
Test name,
Test date,
Test time,
Start load,
Area,
Maximum load,
Maximum strength,
Deform
}
