% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matest.R
\name{matestResultAll}
\alias{matestResultAll}
\title{Matest compression machine test results}
\usage{
matestResultAll(dir)
}
\arguments{
\item{file}{A file name of a text file where the test data are saved.}
}
\value{
A data frame containing test results - raw data.
}
\description{
This function function matestResult to all the files in a directory and all the subdirectories.
}
