% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matest.R
\name{matestTestParameters}
\alias{matestTestParameters}
\title{Matest compression machine test parameters}
\usage{
matestTestParameters(file)
}
\arguments{
\item{file}{A file name of a text file where the test data are saved.}
}
\value{
A data frame containing test parameters.
}
\description{
This function returns parameters of a tested specimen such as:
"Test",
"Test date",
"Test time",
"Maximum load",
"Axial gage length",
"Lateral gage length",
"Area",
"Axial youngs modulus",
"Poissons ratio"
}
