% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{paraugu.kopas}
\alias{paraugu.kopas}
\title{Shows samples in a folder}
\usage{
paraugu.kopas(dir_path)
}
\arguments{
\item{dir_path}{A path to the folder where the raw data are stored}
}
\value{
A data frame containing list of samples
}
\description{
This function returns a data frame showing samples in the specified directory.
The sample name is the corresponding file name.
}
