% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plot.diagram.peaks}
\alias{plot.diagram.peaks}
\title{Plots diagram with peak values}
\usage{
\method{plot}{diagram.peaks}(x1, x2, peaks1, peaks2, cols1 = c(1, 2), cols2 = c(1, 2))
}
\arguments{
\item{x1, x2}{data frames containing measured data that needs to be compared}

\item{peaks1, peaks2}{data frames containing peak values of x1 and x2}

\item{cols1, cols2}{column names of the data frames - c("Time", "Value") excpected
should be the same in x1 and peaks1 and in x2 and peaks2}
}
\description{
This function plots a diagram consisting of two curves and loacal peak values
to be compared.
}
