% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rissFox.R
\name{rissFoxDataFrame}
\alias{rissFoxDataFrame}
\title{Nolasa datus no RissFox Mini un pārveido par 'data frame' tabulu}
\usage{
rissFoxDataFrame(fails)
}
\arguments{
\item{fails}{Ceļš līdz failam un faila nosaukums ar paplašinājumu.
Svarīgi, ka tas ir txt fails. Jo programma eksportē kā teksta failu, bet
kļūdaini paplašinājumu uzdod xls. Tāpēc šis paplašinājums katram failam ir
jāpārmaina uz txt.}
}
\value{
Tabula 'data frame' formātā ar atbilstošiem datu formātiem.
}
\description{
Šī funkcija paņem datus, kas eksportēti no RissFox Mini data logger.
Kolonnu nosaukumi tiek nolasīti no izejas faila.
Šī funkcija veidota RissFox iekārtām (Mini un Multisensor), kas iegādātas
20... gadā ... (serijas Nr., modelis ..) un vadītas ar programmu ...
}
