% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rissFoxPlot.R
\name{rissFoxPlot}
\alias{rissFoxPlot}
\title{Parāda grafiku visiem kanāliem}
\usage{
rissFoxPlot(x)
}
\arguments{
\item{x}{Dati 'data frame' formātā}
}
\value{
ggplot plot
}
\description{
Šī funkcija parāda grafisko attēlojumu visiem RissFox kanāliem pēc laika
}
