% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atena.R
\name{tuvinajumi.saraksts}
\alias{tuvinajumi.saraksts}
\title{Returns list of monitor files}
\usage{
tuvinajumi.saraksts(dir_path)
}
\arguments{
\item{dir_path}{a folder containing monitor files}
}
\description{
This function returns list of monitor files in a folder.
}
