% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebPlotDigitizer.R
\name{wpdDataFrames.xy}
\alias{wpdDataFrames.xy}
\title{Converts WebPlotDigitizer csv file to list of dataframes}
\usage{
wpdDataFrames.xy(csv_file)
}
\arguments{
\item{csv_file}{a csv file exported form WebPlotDigitizer}
}
\value{
list of data frames
}
\description{
This function converts csv files obtained from WebPlotDigitizer to list of data frames.
It assumes that the file consists of X and Y coordinates - two columns for each data frame.
}
