% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{calc.Ecm}
\alias{calc.Ecm}
\title{Modulus of elasticity of concrete}
\usage{
calc.Ecm(fck, kE = 9500)
}
\arguments{
\item{fck}{concrete compressive strength (characteristic value).}

\item{kE}{coefficient considering type of aggregates.}
}
\value{
modulus of elasticity Ecm
}
\description{
This function calculates modulus of elasticity of concrete according to EC2 (2023) (5.1).
}
